Change how configuration is installed, add ImageMagick policy to fix pdf conversion
