+++ electronics/scripts/config/eeschema
